lets add some limitations to the application for using ai. We will have subscription and single time purchaces later, but lets make the code base and db ready for this.
Limit the number of image generation to 3 for each day for each user or session. Limit the number of travel plans to 10. Limit having AI suggestions to 10 for outfit and activity.

add user premium level in db to increase limitations later for premium users. Not auth users will use free tier.